requirements.txt
nltk==3.5
tensorflow==1.9.0
h5py==2.8.0
Sastrawi==1.0.1
numpy==1.16.0
tflearn==0.3.2
scipy==1.2.1
grpcio==1.29.0
grpcio-tools==1.29.0
